refactor: MemberCountField 분리
